+++ casper-schedule/src/main/kotlin/hs/kr/entrydsm/schedule/domain/schedule/application/service/UpdateSchedulesService.kt
@@ -2,8 +2,8 @@ package hs.kr.entrydsm.schedule.domain.schedule.application.service
 
 import hs.kr.entrydsm.schedule.domain.schedule.adapter.`in`.dto.ScheduleDto
 import hs.kr.entrydsm.schedule.domain.schedule.adapter.`in`.dto.request.UpdateSchedulesRequest
-import hs.kr.entrydsm.schedule.domain.schedule.application.exception.InvalidScheduleRequestException
-import hs.kr.entrydsm.schedule.domain.schedule.application.exception.InvalidScheduleSequenceException
+import hs.kr.entrydsm.schedule.domain.schedule.exception.InvalidScheduleRequestException
+import hs.kr.entrydsm.schedule.domain.schedule.exception.InvalidScheduleSequenceException
 import hs.kr.entrydsm.schedule.domain.schedule.application.port.`in`.UpdateSchedulesUseCase
 import hs.kr.entrydsm.schedule.domain.schedule.application.port.out.SaveSchedulePort
 import hs.kr.entrydsm.schedule.domain.schedule.facade.ScheduleFacade
