+++ casper-schedule/src/main/kotlin/hs/kr/entrydsm/schedule/domain/schedule/exception/InvalidScheduleSequenceException.kt
@@ -1,4 +1,4 @@
-package hs.kr.entrydsm.schedule.domain.schedule.application.exception
+package hs.kr.entrydsm.schedule.domain.schedule.exception
 
 import hs.kr.entrydsm.schedule.global.error.exception.CasperException
 import hs.kr.entrydsm.schedule.global.error.exception.ErrorCode
